Increase robustness of text detection
I was able to make the text detection more accurate by checking if each
cropped image has text that fuzzily matches one of the units that would
be expected for the local axis. This slows down the program quite signi-
ficantly however I do not doubt that this can be improved upon. I also
slightly improved the way that vertically oriented text is dealt with, by
simply rotating any cropped images that have h > 2*w. This will obviously
not work in every case however it addresses the issue for the time being.

There are some remaining issues with processing graphs that have two rows
of units, this however should be a trivial fix as it just requires sepa-
rating the cropped images that are all inline with each other into sepa-
rate lists.

There is the overarching issue of the lack of accuracy associated with
OCR. As it does not seem to be an effective strategy to attempt to train
a tesseract specifically for the text styles used in the many types of
fan curves, I believe that the best way to deal with errors is to become
less fragile. One idea I have is to create a running linear scale based
on the coordinates of each cropped image and check if each read value fits
the trend. If anomalies are detected, they would be removed and the rest
of the values can be interpolated. Something like this will eventually
become necessary as I approach the goal of this project: generating fan
curves that overlay system curves for any given fan.

In terms of scraping Octopart for above-mentioned fan curves, I have not
made a huge amount of progress. It seems that it may be quite difficult
to extract data from pdfs with the libraries (fitz) that I have tried
using. When graphs are included in a pdf as an image, fitz can find them,
however it tends to invert the image for some reason. When the graph is
"integrated" into the pdf (ie, the axis data of the graph is copiable),
fitz seems to be unable to extract anything other than the axis values.
This leads me to believe that in order for this project to work, OCR is
an absolute necessity.

My plan for extracting graph images from any given pdf is by automatically
screenshotting every datasheet listed on Octopart and then searching
through the screenshots for fan curves. The second part here proves to be
quite tricky. It seems like opencv does have a method for matching a
'subimage' to some given image, however I only had success with this if
I gave the method the exact graph that was on the page. If I want this to
work for pdfs where I don't already have an image of the graph that may
or may not be located on the page, I may need more powerful tools.