Switch from taking screenshots of pdfs to using pdf2img library
This library works quite nicely, there was a bit of trickiness with dow-
nloading the poppler library and saving it in a location that pdf2img is
aware of, although nothing too crazy.

Now the next tasks in the fan curve finding portion of the project are
a) using the octopart api to save all of the datasheets listed for some
set of parameters
b) searching an image of a pdf for a graph and cropping that graph out
of the pdf.